+++ client/src/game/shapes/line.ts
@@ -70,8 +70,9 @@ export class Line extends Shape {
     snapToGrid(): void {}
     // eslint-disable-next-line @typescript-eslint/no-empty-function
     resizeToGrid(): void {}
-    resize(resizePoint: number, point: GlobalPoint): void {
+    resize(resizePoint: number, point: GlobalPoint): number {
         if (resizePoint === 0) this.refPoint = point;
         else this.endPoint = point;
+        return resizePoint;
     }
 }